back_end/requirements.txt
django
daphne
djangorestframework
channels
django-cors-headers
openai
dotenv
psycopg2-binary
requests
sqlparse
asgiref
certifi
oauthlib
googletrans
pillow